+++ dio/CHANGELOG.md
@@ -11,6 +11,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
 - `IOHttpClientAdapter.onHttpClientCreate` has been deprecated and is scheduled for removal in
   Dio 6.0.0 - Please use the replacement `IOHttpClientAdapter.createHttpClient` instead.
 - Using `CancelToken` no longer closes and re-creates `HttpClient` for each request when `IOHttpClientAdapter` is used. 
+- Improve performance when sending binary data (`List<int>`/`Uint8List`). 
 
 ## 5.1.2